sparse-checkout: list folders in cone mode
When core.sparseCheckoutCone is enabled, the 'git sparse-checkout set'
command takes a list of folders as input, then creates an ordered
list of sparse-checkout patterns such that those folders are
recursively included and all sibling entries along the parent folders
are also included. Listing the patterns is less user-friendly than the
folders themselves.

In cone mode, and as long as the patterns match the expected cone-mode
pattern types, change the output of 'git sparse-checkout list' to only
show the folders that created the patterns.

With this change, the following piped commands would not change the
working directory:

	git sparse-checkout list | git sparse-checkout set --stdin

The only time this would not work is if core.sparseCheckoutCone is
true, but the sparse-checkout file contains patterns that do not
match the expected pattern types for cone mode.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>